got games somewhat working with a work around
experimented with a lot of things. could get the html to be recognized but jekyll would refuse to build the javascript unless I did everything as a link instead of using includes.
Will draw up some boiler plate markdown for the pages and then hyperlink to the html games best execution i can think of without tearing jekyll apart